+++ attributes/default.rb
@@ -1,27 +1,21 @@
-# An array of project names, uris and their branches need to build the NativeX snowflake project 
-default['snowflake-nativex']['snowflake_git_dependencies'] = [{:name => 'twitter-scala-parent-overrides',
-													          :uri => 'git@github.com:nativex/twitter-scala-parent-overrides.git',
-													          :branch => 'master',
-													          :depth => 1},
-													          {:name => 'apache-scribe-client-overrides',
-													          :uri => 'git@github.com:nativex/apache-scribe-client-overrides.git',
-													          :branch => 'master',
-													          :depth => 1}]
 # The NativeX snowflake project name
-default['snowflake-nativex']['nativex_snowflake_project_name'] = 'snowflake'
-# The snowflake git repository uri
-default['snowflake-nativex']['snowflake_git_repository_uri'] = 'git@github.com:nativex/snowflake.git'
-# The snowflake revision or branch you want to checkout.
-default['snowflake-nativex']['snowflake_git_repository_branch'] = 'master'
-# The depth you want to which you want to clone the snowflake project.
-default['snowflake-nativex']['snowflake_git_clone_depth'] = 1
-# The operations home directory where github.com authentication will be setup
-default['snowflake-nativex']['ssh']['home'] = '/root/.ssh'
-# The user that owns the ssh keys.
-default['snowflake-nativex']['ssh']['user'] = 'root'
-# The group that owns the ssh key.
-default['snowflake-nativex']['ssh']['group'] = 'root'
-# The destination directory where the snowflake project will live after being compiled.
-default['snowflake-nativex']['link']['destination_directory'] = '/usr/local'
+default['snowflake-nativex']['application_name'] = 'snowflake'
+# List of tarballs for installing snowflake from archive.
+default['snowflake-nativex']['archive'] = [{:name => 'twitter-scala-parent-overrides',
+												:url => 'https://s3-us-west-2.amazonaws.com/archive-code-nativex/twitter-scala-parent-overrides.tgz'},
+								    			{:name => 'apache-scribe-client-overrides',
+								    			:url => 'https://s3-us-west-2.amazonaws.com/archive-code-nativex/apache-scribe-client-overrides.tgz'},
+								    			{:name => 'snowflake',
+								    			:url => 'https://s3-us-west-2.amazonaws.com/archive-code-nativex/snowflake.tgz'}]
+# Snowflake install method. Currently supports 'archive' and 'source'.
+default['snowflake-nativex']['install_method'] = 'archive'
 # The snowflake project environment home directory
 default['snowflake-nativex']['snowflake_home'] = '/usr/local/snowflake'
+# Snowflake logging file name.
+default['snowflake-nativex']['snowflake_log'] = 'snowflake.log'
+# ELB name ** Must be deployed prior to snowflake instances. **
+default['snowflake-nativex']['elb']['name'] = "DAW1AL-flakeELB"
+# The destination directory where the snowflake project will live after being compiled.
+default['snowflake-nativex']['link']['destination_directory'] = '/usr/local'
+# Zookeeper hosts lists. ##TODO Make this dynamic with an LWRP. *zookeeper cookbook dependency
+default['snowflake-nativex']['zookeeper']['host_list'] = 'pchdvl-zookpr01.teamfreeze.com,pchdvl-zookpr02.teamfreeze.com,pchdvl-zookpr03.teamfreeze.com'