+++ metadata.rb
@@ -4,7 +4,9 @@ maintainer_email 'adrian.herrera@nativex.com'
 license          'All rights reserved'
 description      'Installs/Configures snowflake-nativex'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '0.1.0'
+version          '1.0.0'
 
+depends 'ark'
+depends 'aws'
 depends 'git'
 depends 'maven'