+++ CHANGELOG.md
@@ -12,6 +12,14 @@ This file is used to list changes made in each version of the snowflake-nativex 
 	- links the compiled snowflake project to /usr/local/
 	- creates a snowflakerc file to export the snowflake home directory.
 
+1.0.0
+-----
+-[Adrian Herrera] - Production release of snowflake-nativex
+   - adds support for install from source and archive.
+   - adds dynamic datacenter and worker id configuration for snowflake project.
+   - adds init.d script for stop, start and restart support of application.
+   - adds registration with ELB support.
+
 
 - - -
 Check the [Markdown Syntax Guide](http://daringfireball.net/projects/markdown/syntax) for help with Markdown.